Add staffModel
On branch pauls-api-master
Changes to be committed:
	modified:   controllers/staffController.js
	new file:   models/staffModel.js
